+++ showcase/integrations/langgraph-python/src/agents/beautiful_chat.py
@@ -33,6 +33,8 @@ from langchain_core.tools import tool as lc_tool
 from langchain_openai import ChatOpenAI
 from langgraph.types import Command
 
+from src.agents._a2ui_utils import has_root_component, sanitize_a2ui_components
+
 
 # ─── Shared state schema ────────────────────────────────────────────
 
@@ -269,7 +271,11 @@ def generate_a2ui(runtime: ToolRuntime[Any]) -> str:
 
     prompt = f"{_GENERATE_A2UI_PROMPT_HEADER}\n\n{context_text}".strip()
 
-    model = ChatOpenAI(model="gpt-5.4")
+    # `streaming=True` so aimock's record/replay (which only intercepts
+    # SSE streams) sees this secondary LLM call. Without it the call
+    # bypasses fixture matching in replay mode, surfacing as
+    # "An internal error occurred" on the demo page. Mirrors a2ui_dynamic.py.
+    model = ChatOpenAI(model="gpt-5.4", streaming=True)
     model_with_tool = model.bind_tools(
         [_design_a2ui_surface],
         tool_choice="_design_a2ui_surface",
@@ -289,18 +295,10 @@ def generate_a2ui(runtime: ToolRuntime[Any]) -> str:
     # Force the canonical catalog ID — the secondary LLM has been observed
     # hallucinating IDs from sibling demos when context is sparse.
     catalog_id = CUSTOM_CATALOG_ID
-    components = args.get("components", [])
+    components = sanitize_a2ui_components(args.get("components", []))
     data = args.get("data", {})
 
-    # Defensive: every component must have an `id` AND a `component` field, or
-    # the frontend renderer throws "Cannot create component <id> without a
-    # type" and never recovers. Drop malformed entries so a valid surface
-    # still renders.
-    components = [
-        c for c in components
-        if isinstance(c, dict) and c.get("id") and c.get("component")
-    ]
-    if not any(c.get("id") == "root" for c in components):
+    if not has_root_component(components):
         return json.dumps({
             "error": "LLM produced no valid root component for the A2UI surface."
         })
