--- showcase/integrations/langgraph-python/src/app/demos/tool-rendering/parse-json-result.ts
@@ -1,8 +1,0 @@
-export function parseJsonResult<T>(result: unknown): T {
-  if (!result) return {} as T;
-  try {
-    return (typeof result === "string" ? JSON.parse(result) : result) as T;
-  } catch {
-    return {} as T;
-  }
-}
